impl bfs based on bfs.cpp, passed twitter
one big note:
with cpp , the priority_queue naturally uses
failling order for elements, -- 4 3 2 1
and bfs.cpp utilizes this feature,
but in java, the PriorityQueue naturally uses
raising order for elements, -- 1 2 3 4

notes that breath first search with a bound
requires the closest possible node to be
searched then others. as done in bfs.cpp,
DIST[u] = k (bound) - distance to the root.
the identifier `distance' in the source is
quite misleading, it should be a departure
number from the root.

for example(real bug):
        root=1, bound=3;
        1-->2
        2-->5
        1-->5
        5-->4

when we are computing DIST for node(4)
we already knew DIST[5] == 1 from node(1),
the root, and DIST[5] == 0 from node(2),
the neighbor of the root.
so which value should we use to compute DIST[4]?
it should be DIST[1] from node(1).
otherwise, the node(4) will be left out, cause
DIST[5] reaches zero and DIST[4]<0, 4 is considered
out of bound 2.

when comparing to sssp, bfs is in fact the same as it.
except the edge weight is all one and taken a `bound minus dist'
trick that requires a failling comparator in java.