impl cc, tested on simple, naive,
this algorithm has several things to
notice when doing impl and tesing:

1.the graph is UNDIRECTED though the input file
has no difference with the directed graph.
so the edge `1 2' means (1)<-->(2)
we must consider both directions when treating
the neighbors of a node.
if we visit outgoing, then incoming must follow.
so do parents and children (in this case)

2.the usages of 3 maping maintained by cc:
<1> localComp: maps vid to localCompID.
<2> globalComp: maps localCompID to globalCompID.
<3> outerNodes: maps localCompID to the nodes
linked to localCompID but no contained by this
fragment.

further explanation:
1.
PEval computes <1> and this maping won't
change ever since. localCompID takes the smallest
vid among this component, but not limited to
the parts of node that reside in this fragment.
for example:

        1 <--> 2

where fid(1)=0, fid(2)=1, the localCompID for 1
is 1 of course, but localCompID(2) is 1 too.
because smaller is better, and 1 is visible to 2
via frament.getglobal().getnode(1).

In fact, PEval on each fragment not only computes their
local components i.e. localCompID(vid) for inner nodes,
but also localCompID for nodes that live closest to
the boundaries of each fragment, yet not belong to that
fragment. these information is kept in the outerNodes
.i.e, <3> maping. the interesting fact is that, `outerNodes`
are grouped by their localCompID, so later in IncEval, we
can trace these changes brought by IN-MESSAGE via localCompID.
that is:

IN-MESSAGE == <vid, globalCompID>

<vid, localCompID> ==> <localCompID, Array[outerNodes]>

and thus get their fids so that OUT-MESSAGE can be sent.

2.
IncEval is responsible to minimize the globalCompID
of each node. on receiving IN-MESSAGE, it does:

<step1>
gets localCompID of vid from a piece of message <vid, globalCompID>
updates the old-globalCompID indexed by localCompID.

<step2>
and records the updated localCompID.
notes that it doesn't record vid because outerNodes is not indexed
by vid, nor will it be possible to has vid as index, since
the vids of a component is not unique.
then, IncEval makes a reduction on the maping of
<updated localCompID, Array[ outerNodes ]>
based on the follow:

1.smaller globalCompID for `updated localCompID` is preferred.
(via maping <2>)
2.no duplicated vids of outerNodes should be selected.

and the reduction produce `messages`, whose contents are sent
by getting fid from fragment.getglobal().getnode(vid).getfid()
and the updated globalCompID as value(done in step 1)

in the end, assemble just pulls up all the globalCompID for
all vid. following 2 holds:

1. nodes in different components have different Component ID.
2. the Component ID of a component is the smallest vid among its nodes.