impl simquery and multi-thread version
add one cmd arg, a boolean, to indicate
if we use multi-thread. inside worker added a nested class
workerthread, which extends Thread and impl run.
Also, worker has one more field to record the threads that
it creates, so as to join them. the creation of threads is
done in `sandbox`